feat: Add initial Sidebar layout (incomplete)
