+++ src/components/layout/sidebar/SideMenuItem.tsx
@@ -1,0 +1,10 @@
+import { useDashboardStore } from "@/lib/store/useDashboardStore";
+
+export default function SideMenuItem() {
+  const dashboardId = useDashboardStore((state) => state.dashboardId);
+  return (
+    <div>
+      <div>카드</div>
+    </div>
+  );
+}
