+++ src/components/layout/sidebar/SideMenuList.tsx
@@ -1,0 +1,9 @@
+import SideMenuItem from "./SideMenuItem";
+
+export default function DashboardList() {
+  return (
+    <div>
+      <SideMenuItem />
+    </div>
+  );
+}
